refactor: rename applyPreset to preInitialization
