+++ packages/cli/src/index.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 // MUST import first to apply preset from args and set ssz hasher
-import "./applyPreset.js";
+import "./preInitialization.js";
 import {YargsError} from "./util/index.js";
 import {getLodestarCli, yarg} from "./cli.js";
 import "source-map-support/register.js";
