+++ packages/cli/src/preInitialization.ts
